drop edit-distance graph correction, keep cell barcode selection
